### 2019-05-17 #1456 @arfon
:wave: @adswa, thanks for your submission to JOSS. While I identify the right editor for this submission, could you take a look at [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few individuals that might be able to review this submission?
