### 2019-05-17 6f39b60
remove superflous reference to repo (linked anyway)
