+++ paper.bib
@@ -22,14 +22,6 @@
   url={"https://doi.org/10.3758/s13428-012-0212-2"}
 }
 
-@online{multimatch_gaze,
-    author = {Adina Wagner},
-    title = {multimatch_gaze},
-    year = 2019,
-    howpublished = {https://github.com/adswa/multimatch_gaze},
-    urldate = {2019-04-18}
-}
-
 @article{Johansson,
     author = {Johansson, Roger and Holsanova, Jana and Holmqvist, Kenneth},
     title = {Pictures and spoken descriptions elicit similar eye movements during mental imagery, both in light and in complete darkness},
