### 2019-05-25 #1456 @kyleniemeyer
Hi @FelixHenninger, @oesteban, and @vsoch : are you interested and available to review this submission for JOSS?
