### 2019-05-25 #1456 @vsoch
I would default to @FelixHenninger first- he bought his own timing equipment and published on it! I’ve built infrastructure for deploying experiments, but that’s different from critiquing timing itself. @FelixHenninger do you have the bandwidth?
