### 2019-06-10 #1456 @FelixHenninger
Hi y'all, thanks so much for considering me!

Somehow, I missed the GitHub pings; thanks, @kyleniemeyer, for writing directly! I would be thrilled to take a closer look at this, though I'll be out of office until the end of next week, if it's ok that I complete the review in the week of the 24th, I'd be more than happy to -- the project looks fantastic! If any of you get around to it faster, of course, I totally understand. Kind regards, and sorry for the delay,

-Felix
