### 2019-06-24 #1456 @adswa
Thanks @kyleniemeyer for keeping us up to date - this transparency and communication is highly appreciated!
