### 2019-06-25 #1525 @whedon
[REVIEW]: multimatch_gaze: The MultiMatch algorithm for gaze path comparison in Python
**Submitting author:** @AdinaWagner (<a href="http://orcid.org/0000-0003-2917-3450">Adina Wagner</a>)
**Repository:** <a href="https://github.com/adswa/multimatch_gaze" target ="_blank">https://github.com/adswa/multimatch_gaze</a>
**Version:** v0.1.1
