### 2019-07-05 #1525 @JonathanReardon
Hello, (1) could the list of dependencies be added to the README.md? It states that all will be installed with pip, but it might be better to explicitly state what those dependencies are. (2) Could the formatting of the top (author/affiliation etc.) section of paper.md be improved for clarity?
