### 2019-07-05 f742715
add dependency information to README.md
