### 2019-07-05 #1525 @JonathanReardon
The documentation is very good, and the "Example computation" is particularly useful (https://multimatch.readthedocs.io/en/latest/example.html). Do you state anywhere what you used to overlay the data onto movie frame image (Figure 1 and 2)? if not, I think it would be a useful information to add.
