### 2019-07-05 77863b6
extended contributing/support section.
Link to CONTRIBUTING.md
