### 2019-07-05 7558ba9
do not scale figures, use figwidth instead
This doesn't distort figures when viewing from e.g. mobile devices
