### 2019-07-05 161d17f
Add reference to studyforrest overlay_gaze_on_video.py script
