+++ docs/source/example.rst
@@ -96,6 +96,8 @@ duration was 4.39 seconds (mean = 4.36s).
 The following figures give an overview of the similarity computations.
 Figures 1 and 2 display a frame within the segments in the first run of the movie
 with the lowest and highest group-level similarity (averaged across the five dimensions).
+The overlayed eye gaze was created with a custom script that is part of the studyforrest
+phase-2 publication and publicly available in the corresponding Github repository_.
 
  .. figure:: ../img/low_sim.png
    :figwidth: 100%
@@ -135,6 +137,9 @@ deviations are very small. This is also highlighted by Figure 3.
 
    Distribution of similarity measures throughout the movie. Note the extremely high
    position and length dimension.
+
+.. _repository: https://www.github.com/psychoinformatics-de/studyforrest-data-phase2/blob/master/code/overlay_gaze_on_video
+
 
 Discussion
 ^^^^^^^^^^
