### 2019-07-05 #1525 @adswa
Hey @JonathanReardon,
Thanks for these helpful remarks.

The dependencies are now explicitly stated in the README and in the docs.

As for formatting the paper.md header section I'm unfortunately unsure how to do that, at the moment. @kyleniemeyer, is it possible to line-break those affiliations or similar?

Indeed, the "how to contribute" section was only a minor section in the README.md, and I have now added a CONTRIBUTING.md file (with guidelines on contributing, seeking help, and a code of conduct).

I have linked the script used to overlay gaze onto the stimulus content (its publicly available as part of the studyforrest publication) in the example in the docs - I agree that this is helpful information. Also, many thanks for reading the documentation so attentively!

The changes made so far are in [this PR](https://github.com/adswa/multimatch_gaze/pull/30). I'll try to also add a header improvement to it, and will make a note here when I merge them into master.
