### 2019-07-05 #1525 @JonathanReardon
Hi @adswa -- thank you for the swift response and the changes! I have now finished my part of the review (all boxes ticked). I thoroughly checked the tool and I found it very easy to use, it works exactly as stated. Docs/instructions were also thorough and very clear, which I very much appreciate. Note: I tested it on Linux Mint (19.1) -- I'm not sure what systems you've tested it on but it might be useful to know that it works on this distribution too. It's great to see another useful tool re-implemented for the open source world, congrats (to all involved)!

I'm not sure how it works with respect to the other reviewer, but from my end, I'm happy to approve for publication.
