### 2019-07-05 #1525 @kyleniemeyer
@adswa @JonathanReardon if you mean the affiliation info in the PDF linked in the last comment, the format is part of the JOSS template. The Markdown file is just the original source, so don't worry about how things look in that.

(also, one comment for @adswa: could you add more detail to the affiliation info for your colleague Yaroslav? That one should have the same info, like department, city, state, country.)
