### 2019-07-05 #1525 @adswa
Thanks @kyleniemeyer, the affiliation is updated, and I merged all changes in [this commit](https://github.com/adswa/multimatch_gaze/commit/29636ffd0b0d76e45d08970b7652b4a8b4aab14c). No worries, @JonathanReardon! It did not come across in a bad way at all!

Have a good weekend everyone!
