### 2019-07-18 6a8eced
Switched naive dijkstra implementation with the scipy module implementation of dijkstra. This was done in order to address issue #28.
