### 2019-07-18 11eb98e
Removed the naive dijkstra implementation and replaced it with the scipy module implementation of dijkstra. This was done in order to address issue #28.
