### 2019-07-18 1f91adc
Fixed formatting issues to be more consistent with current format
