### 2019-07-24 c6f6b96
Change dijkstra functionality by using numpy arrays to convert weighted graph into csr_matrix. Also updated setup.py and requirements.txt to include scipy module
