### 2019-08-06 0024e91
Changed the implementation of createdirectedgraph in order to better fit the data format for scipy dijkstra algorithm. Updated docomparison_forrest in utils.py file as well in order to fit the new format
