### 2019-08-11 #1525 @FelixHenninger
Hej everyone, and sincerest apologies for the delay, which has been due to me being out of office in in July, followed by health issues on my part over the past ten days -- I'm very sorry for the trouble! My brief notes follow:

This project has been a genuine pleasure to review as the installation and use work flawlessly, and as a result I have only minor points to add to @JonathanReardon's excellent (and already-addressed) comments above. I emphatically agree that the documentation is fantastic, and also found the references describing the package's inner workings very clear and helpful when looking through the functionality. I had a couple of things that were unclear to me at first, but in every case it turned out that I hadn't read the docs carefully enough. 🙈

I found the manuscript very helpful and accessible, with a clear motivation and concise description of the package. I have only two minor formatting-related points that might be worth a second look:

* On lines 54/55, I think the citations might work better in the [in-text format](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax) rather than in brackets.
* In the bibliography, I think that the PDF generation might have garbled the capitalization for Behavior Research Methods. There are also some minor inconsistencies with the DOI formatting, and some of the links don't work as a result (Burmeister & Mast, Dewhurst et al. and Johansson et al.).

There is one minor point I encountered that I think is the scope of the JOSS review, so please consider this as suggestion only (if they aren't covered elsewhere already -- I'd be happy to make an issue out of it if that's a better place for the information):
* It's really great to see CI testing in a project like this! 😍 One of the things I had missed initially is the automated testing, and thus I had tried running the test suite locally, which isn't (as far as I can see) documented -- it might be helpful for future contributors to briefly note how to trigger a test run manually.

Ok, I hope this makes sense -- again, thanks for putting together this super-neat package, and doing it with such care and visible attention to detail. If you, @adswa, could take another look at the manuscript formatting, I think this is a go from my perspective. 👍 Sorry again for the delay!
