### 2019-08-12 3a7eb0e
include pytest instructions
