### 2019-08-12 #1525 @adswa
Hi @FelixHenninger, sorry to hear about health issues - don't worry about the delay. Glad to hear you find it useful, and thanks for your comments. :slightly_smiling_face:

In this [PR](https://github.com/adswa/multimatch_gaze/pull/33) I've added a note about how to run the test suite locally into the README.md, CONTRIBUTING.md, and the docs. I find it to be a very useful addition, thanks for the suggestion!
I also adjusted the manuscript so that the citations you mentioned are now in-text, and fixed the broken dois (thanks for checking so thoroughly). I forgot the BRM capitalization and did this [here](https://github.com/adswa/multimatch_gaze/pull/34) :smile:
