### 2019-08-12 #1525 @adswa
Retrying in order to get the doi with an underscore to work.
