### 2019-08-12 893c02b
minor spelling fixes
