### 2019-08-13 #1525 @kyleniemeyer
@adswa my only editorial request for the paper is to add
> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

(This is from our documentation on [what papers should contain](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).)

This could take multiple forms, but the simplest is probably to add a sentence or two explaining the application to a general reader at the very beginning.
