### 2019-08-13 #1525 @adswa
Thanks a lot, @kyleniemeyer, I've included a short description at the very beginning into the introduction in [this PR](https://github.com/adswa/multimatch_gaze/pull/37):

> ``multimatch-gaze`` is a Python package for computing the
similarity of scan paths from eye gaze data on five different dimensions.

Let me know if that suffices at as a high-level functionality description, or if that is not detailed enough. Thanks!
