### 2019-08-13 #1525 @kyleniemeyer
@adswa this is better, but I think it still needs a bit more general explanation for non-experts—for example, I don't really know what scan paths are or what they are used for. Could you add a sentence explaining that?
