### 2019-08-15 #1525 @adswa
Thanks a lot for the feedback, @kyleniemeyer! I made another attempt in [this PR](https://github.com/adswa/multimatch_gaze/pull/38)
