### 2019-08-15 7d5b0d6
Minor edit to paper.md
man-machine-interface -> human-machine interface
