### 2019-08-15 #1525 @kyleniemeyer
@adswa that looks great! I just made a minor edit in this PR, if you could merge: https://github.com/adswa/multimatch_gaze/pull/39

Once that is complete, please archive the entire repository (e.g., using Zenodo) and let me know the DOI here.
