### 2019-08-16 #1525 @kyleniemeyer
@openjournals/joss-eics ok, this submission is ready to accept!
