### 2019-08-16 #1525 @danielskatz
👋 @kyleniemeyer - I think @openjournals/joss-eics members can do this themselves, even if not on duty - at least @arfon and I have in the past. There's certainly no requirement to, but it's an option if you want to.
